refactor: extract inspection logic
